Fix Access when the only readable ref is a RegExp with a gerrit pattern
When no other ref than a RegExp with a gerrit username or shardeduserid
patthern, The Permission is not properly validated and repository access
is denied.

Change-Id: Id6a890a91c5d6d954dab74694c2d3fd53c15897c
Fix Access when the only readable ref is a RegExp with a gerrit pattern

When no other ref than a RegExp with a gerrit username or shardeduserid
patthern, The Permission is not properly validated and repository access
is denied.

Change-Id: Id6a890a91c5d6d954dab74694c2d3fd53c15897c